Archive the muting system and replace it with discord's timeout feature since it's much easier to handle and use
